lmodifier OK mais a tester en detail

+++ includes/libft/ft_dtoa.c
@@ -6,16 +6,16 @@
 /*   By: jjauzion <jjauzion@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/07 18:25:59 by jjauzion          #+#    #+#             */
-/*   Updated: 2018/01/11 17:40:09 by jjauzion         ###   ########.fr       */
+/*   Updated: 2018/01/17 11:58:39 by jjauzion         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char			*ft_dtoa(double nbr, int precision)
+char			*ft_dtoa(long double nbr, int precision)
 {
 	int			n;
-	double		real;
+	long double	real;
 	char		*mynum;
 	int			len;
 	int			i;
@@ -28,7 +28,7 @@ char			*ft_dtoa(double nbr, int precision)
 	len = ft_strlen(mynum);
 	mynum = (char *)ft_realloc((void **)&mynum, len + 1, precision + 2);
 	mynum[len] = '.';
-	real = ABS((nbr - (double)n));
+	real = ABS((nbr - (long double)n));
 	i = -1;
 	while (++i < precision)
 	{
@@ -37,7 +37,7 @@ char			*ft_dtoa(double nbr, int precision)
 //printf("(int)(real*10) = %d\n", (int)(real * 10));
 		n = (int)(real * 10);
 //printf("n = %d\n", n);
-		real = real * 10 - (double)n;
+		real = real * 10 - (long double)n;
 //printf("real = %f\n", real);
 		mynum[len + i + 1] = n + '0';
 		mynum[len + i + 2] = '\0';
